Added custom header support in "call" subroutine.
Issue-1: Call required a different header in order to use the PATCH
method of updating the properties of an image. Now, there is an
optional argument that can be passed to the call subroutine that will
allow one to specify a hash reference of headers. The only restriction
is that the "x-auth-token" header is not allowed to be modified. I also
updated the POD to reflect this change and added a basic unit test.